avoid rogue 2fa code generations
When a wiki uses rewriting, non-existing files are mapped to doku.php
and interpreted as page names. When using a 2fa provider that transmits
codes (like email) this could lead to sending multiple codes out for
each of these bogus requests. This patch ensures that the 2fa form (and
code sending) is only triggered for document requests as indicated by
the sec-fetch-dest request header.